Use one benchmark binary
