+++ demo/soma_clustering/src/soma_clustering.cc
@@ -14,4 +14,6 @@
 
 #include "soma_clustering.h"
 
-int main(int argc, const char** argv) { return bdm::Simulate(argc, argv); }
+int main(int argc, const char** argv) {
+  return bdm::soma_clustering::Simulate(argc, argv);
+}
