+++ demo/soma_clustering/src/soma_clustering.h
@@ -31,6 +31,7 @@
 #include "validation_criterion.h"
 
 namespace bdm {
+namespace soma_clustering {
 
 enum Substances { kSubstance0, kSubstance1 };
 
@@ -90,6 +91,7 @@ inline int Simulate(int argc, const char** argv) {
   return !crit;
 }
 
+}  // namespace soma_clustering
 }  // namespace bdm
 
 #endif  // DEMO_SOMA_CLUSTERING_H_
