+++ demo/tumor_concept/src/tumor_concept.cc
@@ -14,4 +14,6 @@
 
 #include "tumor_concept.h"
 
-int main(int argc, const char** argv) { return bdm::Simulate(argc, argv); }
+int main(int argc, const char** argv) {
+  return bdm::tumor_concept::Simulate(argc, argv);
+}
