+++ test/benchmark/main.cc
@@ -1,4 +1,3 @@
 #include <benchmark.h>
 
 BENCHMARK_MAIN();
-
